// app/types/json-design.types.ts
export type JsonGradColor = {
  c: string
  p: number
}
// ...
export type JsonRGradPos =
  | "center center"
  | "center right"
  | "center left"
  | "bottom left"
  | "bottom center"
  | "bottom right"
  | "top left"
  | "top center"
  | "top right"

export type JsonAlign =
  | "bottom"
  | "center"
  | "left"
  | "middle"
  | "right"
  | "top"

export type JsonScaleMode =
  | "stretch"
  | "aspect"
  | "mask"
  | "tile"
  | "crop"
  | "userCrop"

export type JsonBackgroundType = "solid" | "rgrad" | "lgrad" | "image" | "none"

export type JsonBackgroundSolid = {
  type: "solid"
  scolor: string
}

export type JsonBackgroundSolidText = JsonBackgroundSolid & {
  feed?: AdStudioLayerFeedPropertyBind
}

export type JsonBackgroundLinearGradient = {
  type: "lgrad"
  gradColors: JsonGradColor[]
  backgroundRotation: number
}

export type JsonBackgroundRadialGradient = {
  type: "rgrad"
  gradColors: JsonGradColor[]
  rgradPos: JsonRGradPos
}

export type JsonBackgroundImage = {
  type: "image"
  originalWidth: number
  originalHeight: number
  url: string
  backgroundRotation?: number
  scaleMode: JsonScaleMode
  contentScale?: number
  horizontalAlign: JsonAlign
  verticalAlign: JsonAlign
  contentOffsetX?: number
  contentOffsetY?: number
  hqUrl?: string
  source?: string
  hash?: string
  licensing?: boolean
}

export type JsonBackgroundNone = {
  type: "none"
}

export interface JsonBackgroundBorder {
  useBorder?: boolean
  borderColor?: string
}

export type JsonBackground =
  | JsonBackgroundNone
  | JsonBackgroundSolid
  | JsonBackgroundLinearGradient
  | JsonBackgroundRadialGradient
  | JsonBackgroundImage
// ...
export type JsonShadow = {
  useShadow: boolean
  hShadow: number
  vShadow: number
  blur: number
  spread: number
  color: string
}

export type JsonBlur = {
  useBlur: boolean
  pixels: number
}

export const JsonFlipTypeArr = [
  "both",
  "horizontal",
  "none",
  "vertical",
] as const

export type JsonFlipType = (typeof JsonFlipTypeArr)[number]

export type JsonTextAlignment = "left" | "right" | "center" | "justify"
// ...
export type JsonFontSettings = {
  fontFamily: string
  fontType: FontType
  fontStyle: FontStyle
  fontPrefix: string | null
  fontWeight: number
  fontFaceUrl: string | null
  fontUrl: string | null
}
// ...
export type JsonBorder = {
  color: string
  weight: number
  radius: number
}

export type FontType = "google" | "custom" | "all" | "bannersnack" | "creatopy"
export type FontStyle = "normal" | "italic" | "regular" | "slanted" | "oblique"
// ...
export type JsonAdjustColor = {
  useAdjustColor: boolean
  brightness: number
  contrast: number
  saturate: number
  hue: number
}